pyproject.toml
[tool.poetry]
name = "strelka-worker"
version = "0.24.03.04"
description = "Strelka's backend Python worker"
authors = [
    "Paul Hutelmyer <Paul.Hutelmyer@Target.com>",
    "Ryan Ohoro <Ryan.Ohoro@Target.com>",
    "Sara Kalupa <Sara.Kalupa@Target.com>"
]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<=3.12"
arc4 = "0.0.4"
beautifulsoup4 = "4.12.2"
boltons = "23.0.0"
boto3 = "1.28.60"
construct = "2.10.68"
cryptography = "^44.0.1"
dncil = "1.0.2"
dnfile = "0.14.1"
docker = "6.1.3"
dotnetfile = "0.2.4"
eml-parser = "1.17.5"
esprima = "4.0.1"
formulas = "1.2.6"
html5lib = "1.1"
inflection = "0.5.1"
jsbeautifier = "1.14.9"
libarchive-c = "5.0"
lief = "0.13.2"
lxml = "4.9.3"
m2crypto = "0.39.0"
nested-lookup = "0.2.25"
numpy = "1.26.0"
olefile = "0.46"
oletools = "0.60.2"
opencv-contrib-python = "4.8.1.78"
opencv-python = "4.8.1.78"
openpyxl = "3.0.10"
opentelemetry-api = "1.17.0"
opentelemetry-exporter-jaeger = "1.17.0"
opentelemetry-exporter-otlp = "1.17.0"
opentelemetry-exporter-otlp-proto-grpc = "1.17.0"
opentelemetry-exporter-otlp-proto-http = "1.17.0"
opentelemetry-sdk = "1.17.0"
pefile = "2023.2.7"
pillow-avif-plugin = "1.4.3"
pillow-heif = "0.16.0"
pgpdump3 = "1.5.2"
py-tlsh = "4.7.2"
pycdlib = "1.14.0"
pycryptodomex = "3.20.0"
pyelftools = "0.29"
pygments = "2.15.0"
pylzma = "0.5.0"
PyMuPDF = "1.26.5"
pytesseract = "0.3.10"
pytest-mock = "3.10.0"
pytest-unordered = "0.5.2"
pytest = "7.2.2"
python-docx = "0.8.11"
python-magic = "0.4.27"
py7zr = "0.20.2"
pytz = ">=2022.7.1"
pyxlsb2 = "0.0.9"
pyyaml = "6.0.1"
pyzbar = "0.1.9"
pyzipper = "0.3.6"
rarfile = "4.1"
redis = "4.5.4"
requests = "^2.32.3"
rpmfile = "1.1.1"
signify = "0.5.1"
ssdeep = "3.4"
tldextract = "3.4.0"
tnefparse = "1.4.0"
validators = "0.35.0"
xlrd2 = "1.3.4"
xlrd = "2.0.1"
xmltodict = "0.13.0"
yara-python = "4.3.1"

# Git dependencies
chaskey = { git = "https://github.com/volexity/chaskey-lts", rev = "2fd80f732dd9422a9e92556758180ceee3b5b4ec" }  # For donut-decryptor
donut-decryptor = { git = "https://github.com/volexity/donut-decryptor", rev = "5a0f9f87094efe6e8d131d3fa4abf23f31123784" } # For donut-decryptor
entropy = { git = "https://github.com/jshlbrd/python-entropy.git", rev = "a49f1addccb88d54115832a6d1982b957a56f329" }
speakeasy-emulator = { git = "https://github.com/mandiant/speakeasy.git", rev = "1cb52a92ab4bae3659b0f8db4ed29f591d932c88" }
zipp = "^3.20.1"
certifi = "^2025.1.31"
setuptools = "^75.8.1"
macaddress = "^2.0.2"
pydantic = "^2.12.4"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
